# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Digital card store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  balance: number;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export interface ProfileUpdate {
  username?: string;
  email?: string;
}

export interface TopupInput {
  /** @minimum 1 */
  amount: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  /** @nullable */
  description?: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface CategoryInput {
  name: string;
  slug: string;
  icon: string;
  description?: string;
  sortOrder?: number;
}

export interface Product {
  id: number;
  name: string;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  featured: boolean;
  stock: number;
  salesCount: number;
  /** @nullable */
  region?: string | null;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  categoryId: number;
  price: number;
  originalPrice?: number;
  description: string;
  imageUrl?: string;
  featured?: boolean;
  region?: string;
}

export interface ProductUpdate {
  name?: string;
  categoryId?: number;
  price?: number;
  originalPrice?: number;
  description?: string;
  imageUrl?: string;
  featured?: boolean;
  region?: string;
}

export interface Card {
  id: number;
  productId: number;
  content: string;
  sold: boolean;
  createdAt: string;
}

export interface CardBatchInput {
  productId: number;
  contents: string[];
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  completed: 'completed',
  refunded: 'refunded',
} as const;

export interface Order {
  id: number;
  /** @nullable */
  userId?: number | null;
  productId: number;
  /** @nullable */
  productName?: string | null;
  quantity: number;
  totalPrice: number;
  status: OrderStatus;
  /** @nullable */
  cardContents?: string | null;
  /** @nullable */
  email?: string | null;
  createdAt: string;
}

export interface OrderInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
  email?: string;
}

export interface AdminStats {
  totalOrders: number;
  totalRevenue: number;
  totalProducts: number;
  totalUsers: number;
  totalCards: number;
  availableCards: number;
  recentOrders: Order[];
}

export type ListProductsParams = {
categoryId?: number;
search?: string;
featured?: boolean;
};

export type ListCardsParams = {
productId: number;
};

export type CreateCards201 = {
  added: number;
};

export type ListAllOrdersParams = {
status?: string;
limit?: number;
};