# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Digital card store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # Categories
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Admin - Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Admin - Update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Admin - Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Products
...
  /cards:
    get:
      operationId: listCards
      tags: [cards]
      summary: Admin - List cards for a product
      parameters:
        - name: productId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cards list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
    post:
      operationId: createCards
      tags: [cards]
      summary: Admin - Add cards to a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardBatchInput"
      responses:
        "201":
          description: Cards added
          content:
            application/json:
              schema:
                type: object
                properties:
                  added:
                    type: integer
                required: [added]
  /cards/{id}:
    delete:
      operationId: deleteCard
      tags: [cards]
      summary: Admin - Delete a card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Orders
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List current user's orders
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Purchase - create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created with card content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Out of stock or insufficient balance
        "401":
          description: Not authenticated
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details (card content revealed)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /admin/orders:
    get:
      operationId: listAllOrders
      tags: [admin]
      summary: Admin - List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  # User profile & balance
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, email, balance, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        balance:
          type: number
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
...
    Category:
      type: object
      required: [id, name, slug, icon, sortOrder, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        description:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
    CategoryInput:
      type: object
      required: [name, slug, icon]
      properties:
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
...
    Card:
      type: object
      required: [id, productId, content, sold, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        content:
          type: string
        sold:
          type: boolean
        createdAt:
          type: string
    CardBatchInput:
      type: object
      required: [productId, contents]
      properties:
        productId:
          type: integer
        contents:
          type: array
          items:
            type: string
    Order:
      type: object
      required: [id, productId, quantity, totalPrice, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        quantity:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, completed, refunded]
        cardContents:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        createdAt:
          type: string
    OrderInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        email:
          type: string
...