# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Digital card store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  /** @nullable */
  userId?: number | null;
  productId: number;
  /** @nullable */
  productName?: string | null;
  quantity: number;
  totalPrice: number;
  status: OrderStatus;
  /** @nullable */
  cardContents?: string | null;
  /** @nullable */
  email?: string | null;
  createdAt: string;
}