# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Digital card store API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
  email?: string;
}