# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Digital card store API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}